refactor(arch): 1) Move lógica de negócio dos controllers para os services 2) A função updateTask alterada para receber o userId e fazer a verificação de permissão
